several improvements

+++ src/javelin/InstallCommand.hx
@@ -12,12 +12,17 @@ import sys.io.Process;
 import haxe.Template;
 import haxe.Resource;
 
-class InstallCommand extends JCommand{
+class InstallCommand extends TestCommand{
 
     public function new(){
         super();
-        addPreRequisite(TestCommand);
+    }
 
+    override public function initialise():Void{
+        super.initialise();
+        #if debug
+            print("initialising InstallCommand");
+        #end
     }
 
     override public function execute() : Void{
@@ -26,7 +31,6 @@ class InstallCommand extends JCommand{
         var runFile : String = null;
 
         if(project.runMain != null && project.runMain != ""){
-            print("compiling " + project.runMain + " in run.n");
             runFile = "run.n";
             var runCompileArgs = ["-neko", runFile, "-main", project.runMain];
             for(runClassPath in project.runClassPaths){
@@ -45,6 +49,9 @@ class InstallCommand extends JCommand{
                 runCompileArgs.push("-resource");
                 runCompileArgs.push(runCompileTimeResource + "@" + runCompileTimeResource);
             }
+            if(debug){
+                runCompileArgs.push("-debug");
+            }
             print("compiling : haxe " + runCompileArgs.join(" "));
             var runCompile = Sys.command("haxe", runCompileArgs);
             if(runCompile != 0){
@@ -55,35 +62,53 @@ class InstallCommand extends JCommand{
 
         }
 
-        var mlibTemplate = new Template(Resource.getString("mlib.mtt"));
-        var mlibContent = mlibTemplate.execute({licenseFile:project.licenseFile,classPaths:project.classPaths,runFile:runFile,resources:project.resources,haxelibOutput:project.haxelibOutput}); //TODO add resources ?
-        var mlibFile = createFile(".mlib");
-        mlibFile.writeString(mlibContent, false);
+        var mlibReturnCode = -1;
+        try{
+            var mlibTemplate = new Template(Resource.getString("mlib.mtt"));
+            var mlibContent = mlibTemplate.execute({licenseFile:project.licenseFile,classPaths:project.classPaths,runFile:runFile,resources:project.resources,haxelibOutput:project.haxelibOutput}); //TODO add resources ?
+            var mlibFile = createFile(".mlib");
+            mlibFile.writeString(mlibContent, false);
 
-        var releaseNotes = console.prompt("release notes", 2);//TODO get rid for local install
 
-        var haxelibTemplate = new Template(Resource.getString("haxelib.json.mtt"));
-        var haxelibContent = haxelibTemplate.execute({
-            name:project.name,
-            url:project.url,
-            license:project.license,
-            tags:printArray(project.tags),
-            description:project.description,
-            releaseNotes:releaseNotes,
-            contributors:printArray(project.contributors),
-            dependencies:printDependencies(project.dependencies),
-        });
-        var haxelibFile = createFile("haxelib.json");
-        haxelibFile.writeString(haxelibContent, false);
-
-        var mlibReturnCode = Sys.command("haxelib", ["run", "mlib", "install"]);
-
-        deleteFiles();
-
-        if(mlibReturnCode != 0){
-            error("=> Failed to install");
+            var haxelibTemplate = new Template(Resource.getString("haxelib.json.mtt"));
+            var haxelibContent = haxelibTemplate.execute({
+                name:project.name,
+                url:project.url,
+                license:project.license,
+                tags:printArray(project.tags),
+                description:project.description,
+                releaseNotes:project.releaseNotes,
+                contributors:printArray(project.contributors),
+                dependencies:printDependencies(project.dependencies),
+            });
+            var haxelibFile = createFile("haxelib.json");
+            haxelibFile.writeString(haxelibContent, false);
+            mlibReturnCode = runMlib();
+        }catch(e : Dynamic){
+            print(e);
         }
 
+        if(runFile != null){
+            var runNFile = console.dir.resolveFile(runFile);
+            runNFile.deleteFile();
+        }
+        if(doNotDeleteFileOnError && mlibReturnCode != 0){
+            print("keep file for investigation");
+        }else{
+            deleteFiles();
+        }
+
+
+        if(mlibReturnCode != 0){
+            error("=> Failure");
+        }else{
+            print("=> Success");
+        }
+
+    }
+
+    private function runMlib() : Int{
+        return Sys.command("haxelib", ["run", "mlib", "install"]);
     }
 
     private function printArray(arr : Array<String>) : String{
